# cache Tag __hash__ value for a 10% resolve speedup