### Problem

Via profiling, I found that `pip._internal.models.wheel.Wheel.supported()` was taking a while, and that over 2/3rds of its runtime was calling the `__hash__` method on the `Tag` object.

### Solution

- Add `_hash` slot to `Tag` class and return it in `__hash__()`.

### Result
A 6% speedup for the v1 resolver, and a 10% speedup for the v2 resolver when resolving `tensorflow==1.14.0`
```bash
> git checkout master
> rm -rf tmp/ ~/Library/Caches/pip && time pip download -d tmp/ tensorflow==1.14.0 &>/dev/null
10.69s user 3.57s system 78% cpu 18.043 total
> rm -rf tmp/ ~/Library/Caches/pip && time pip download --unstable-feature=resolver -d tmp/ tensorflow==1.14.0 &>/dev/null
15.07s user 3.57s system 85% cpu 21.682 total
> git checkout -
> rm -rf tmp/ ~/Library/Caches/pip && time pip download -d tmp/ tensorflow==1.14.0 &>/dev/null
9.33s user 3.69s system 76% cpu 16.914 total
> rm -rf tmp/ ~/Library/Caches/pip && time pip download --unstable-feature=resolver -d tmp/ tensorflow==1.14.0 &>/dev/null
12.82s user 3.70s system 84% cpu 19.550 total
```