@pfmoore 2020-06-23
As this is a fix to a vendored package, it should be submitted downstream (to the packaging project, in this case).

But good catch! I'd have never expected hashing a tag value to be such a costly operation 🙁