@cosmicexplorer 2020-06-23
I will reopen this PR with the vendoring change when pypa/packaging#316 is merged upstream!